% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_resights}
\alias{plot_resights}
\title{Visualise recoveries}
\usage{
plot_resights(
  ring,
  data,
  point_size = 4,
  line_width = 0.75,
  limits = NULL,
  scale = c("medium", "large")
)
}
\arguments{
\item{ring}{Ring Nr}

\item{data}{Already-loaded list returned by \code{read_gpkg()}}

\item{point_size}{numeric}

\item{line_width}{numeric}

\item{limits}{Optional. Vector der Form c(xmin, xmax, ymin, ymax)}

\item{scale}{c("medium", "large")}
}
\description{
Visualise recoveries
}
