% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ring_history}
\alias{ring_history}
\title{Summaries captures (ringed and recovered) by day}
\usage{
ring_history(
  gpkg,
  project = "IMS",
  years = NULL,
  min.count = 1,
  xlab = "Datum",
  ylab = "Tagessume"
)
}
\arguments{
\item{gpkg}{path to a RING2.1.gpkg file}

\item{project}{Projectname. defaults to 'IMS'}

\item{years}{optional. vector of years (e.g. 2023:2026)}

\item{min.count}{minimum counter per day. defaults to 1}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\description{
Summaries captures (ringed and recovered) by day
}
